pyproject.toml
[tool.poetry]
name = "de-kandar"
version = "0.1.0"
description = "Select an area on monitor display and play it on tkinter application."
authors = ["Joseph Lim <josephlim_94@hotmail.co.uk>"]
license = "MIT"
readme = "README.md"
packages = [{include = "de_kandar"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.1"
pillow = "^10.0.0"
ffmpeg-python = "^0.2.0"
av = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
